requirements_gpu.txt
numpy==1.14.5
pandas==0.23.3
scipy==1.1.0
tensorflow-gpu==1.10.0
flask==1.0.2